Created a new command line option named common-labels.
With this option, the list of labes provided is read and parsed into a hashmap.

It's been created a new value parser that split the initial string into a list of keypairs. The list is also splitted into key and value, to populate a hashmap. It's been used the regex library for more flexibility.