Refactored code and included common labels.
Now, instead of using global variables, the list of metrics used in the code is gathered in the same struct, which is passed between the different functions that may require it. This way, adding costant labels to the metrics is straightforward, they are read from the command line and included into the telemetry object.

This link refers to a stack overflow post that recommend to avoid global variables in general. https://stackoverflow.com/a/27826181